chore(generative-ai): clean up generative_ai folder
* move files to text_models
* remove empty folder
