--- generative_ai/count_token/requirements-test.txt
@@ -1,4 +1,0 @@
-backoff==2.2.1
-google-api-core==2.19.0
-pytest==8.2.0
-pytest-asyncio==0.23.6
