--- generative_ai/count_token/requirements.txt
@@ -1,14 +1,0 @@
-pandas==1.3.5; python_version == '3.7'
-pandas==2.0.3; python_version == '3.8'
-pandas==2.1.4; python_version > '3.8'
-pillow==10.3.0; python_version < '3.8'
-pillow==10.3.0; python_version >= '3.8'
-google-cloud-aiplatform[all]==1.69.0
-sentencepiece==0.2.0
-google-auth==2.29.0
-anthropic[vertex]==0.28.0
-langchain-core==0.2.11
-langchain-google-vertexai==1.0.6
-numpy<2
-openai==1.30.5
-immutabledict==4.2.0
