+++ generative_ai/text_models/code_completion_test_function.py
