+++ generative_ai/text_models/requirements.txt
