+++ generative_ai/text_models/streaming_chat.py
